Update confirmation.html
Updated the content following a discussion on Slack. Went for the MVP version because all the details of the applications will be sent in an email AND at some point we will have a past applications screen where a user can also find the full details.

Still needs the contact email. And maybe a link to a feedback screen.